+++ packages/dev/core/src/Engines/AbstractEngine/index.ts
@@ -1,2 +1,8 @@
 /* eslint-disable import/export */
 export * from "./abstractEngine.cubeTexture";
+export * from "./abstractEngine.loadingScreen";
+export * from "./abstractEngine.dom";
+export * from "./abstractEngine.states";
+export * from "./abstractEngine.query";
+export * from "./abstractEngine.observable";
+export * from "./abstractEngine.renderPass";
